refactor operations via codex to subdirectory
