+++ backend/services/tools/geoprocess_tools.py
@@ -26,343 +26,18 @@ from langchain_core.messages import HumanMessage
 from services.ai.llm_config import get_llm
 
 
+# Imports of operation functions from geoprocessing ops and utils
+from services.tools.geoprocessing.utils import get_last_human_content
+from services.tools.geoprocessing.ops.buffer import op_buffer
+from services.tools.geoprocessing.ops.centroid import op_centroid
+from services.tools.geoprocessing.ops.simplify import op_simplify
+from services.tools.geoprocessing.ops.overlay import op_overlay
+from services.tools.geoprocessing.ops.merge import op_merge
+from services.tools.geoprocessing.ops.sjoin import op_sjoin
+from services.tools.geoprocessing.ops.sjoin_nearest import op_sjoin_nearest
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
-
-
-def get_last_human_content(messages):
-    for msg in reversed(messages):
-        if isinstance(msg, HumanMessage):
-            return msg.content
-    raise ValueError("No human message found in context")
-
-
-def _flatten_features(layers):
-    """
-    Given a list of Feature or FeatureCollection dicts, return a flat list of Feature dicts.
-    """
-    feats = []
-    for layer in layers:
-        if layer.get("type") == "FeatureCollection":
-            feats.extend(layer.get("features", []))
-        elif layer.get("type") == "Feature":
-            feats.append(layer)
-        else:
-            # skip anything that is not a Feature or FeatureCollection
-            logger.debug(f"Skipping invalid GeoJSON layer: {layer}")
-    return feats
-
-
-def _get_layer_geoms(layers):
-    """
-    Given a list of GeoJSON Feature or FeatureCollection dicts, return a
-    list of Shapely geometries, each being the unary_union of one layer's features.
-    """
-    geoms: List[Any] = []
-    for layer in layers:
-        layer_type = layer.get("type")
-        if layer_type == "FeatureCollection":
-            feats = layer.get("features", [])
-        elif layer_type == "Feature":
-            feats = [layer]
-        else:
-            logger.debug(f"Skipping layer with missing or invalid type: {layer_type}")
-            continue
-
-        if not feats:
-            continue
-
-        try:
-            gdf = gpd.GeoDataFrame.from_features(feats)
-            gdf.set_crs("EPSG:4326", inplace=True)
-            geoms.append(gdf.unary_union)
-        except Exception:
-            logger.exception("Failed to convert features to GeoDataFrame")
-    return geoms
-
-
-def op_buffer(layers, radius=10000, buffer_crs="EPSG:3857", radius_unit="meters"):
-    """
-    Buffers features of a single input layer item individually.
-    If multiple layers are provided, this function will raise a ValueError.
-    Input geometries are assumed in EPSG:4326. This function:
-      1) Expects `layers` to be a list containing a single layer item (FeatureCollection or Feature).
-      2) Converts radius to meters based on radius_unit (default is "meters").
-      3) Extracts features from the single layer item.
-      4) Creates a GeoDataFrame from these features.
-      5) Reprojects the GeoDataFrame to buffer_crs (default EPSG:3857, which uses meters).
-      6) Applies buffer to each feature geometry with the meter-based radius.
-      7) Reprojects the GeoDataFrame (with buffered features) back to EPSG:4326.
-      8) Returns a list containing one FeatureCollection with the individually buffered features.
-    Supported radius_unit: "meters", "kilometers", "miles".
-    """
-    if not layers:
-        logger.warning("op_buffer called with no layers")
-        return []  # No input layer, return empty list
-
-    if len(layers) > 1:
-        # Extract layer names/titles if available for better error information
-        layer_info = []
-        for i, layer in enumerate(layers):
-            name = None
-            if isinstance(layer, dict):
-                props = layer.get("properties", {})
-                if props:
-                    name = props.get("name") or props.get("title")
-                # Also try to get name from features if it's a FeatureCollection
-                if (
-                    not name
-                    and layer.get("type") == "FeatureCollection"
-                    and layer.get("features")
-                ):
-                    first_feat = layer["features"][0] if layer["features"] else None
-                    if first_feat and isinstance(first_feat, dict):
-                        props = first_feat.get("properties", {})
-                        if props:
-                            name = props.get("name") or props.get("title")
-            layer_info.append(f"Layer {i+1}" + (f": {name}" if name else ""))
-
-        layer_desc = ", ".join(layer_info)
-        raise ValueError(
-            f"Buffer operation error: Only one layer can be buffered at a time. Received {len(layers)} layers: {layer_desc}. Please specify a single target layer."
-        )
-
-    layer_item = layers[0]  # Process the single layer provided
-    unit = radius_unit.lower()
-    factor = {"meters": 1.0, "kilometers": 1000.0, "miles": 1609.34}.get(unit)
-    if factor is None:
-        logger.warning(f"Unknown radius_unit '{radius_unit}', assuming meters")
-        factor = 1.0
-
-    actual_radius_meters = float(radius) * factor
-
-    current_features = []
-    if isinstance(layer_item, dict):
-        if layer_item.get("type") == "FeatureCollection":
-            current_features = layer_item.get("features", [])
-        elif layer_item.get("type") == "Feature":
-            current_features = [layer_item]
-
-    if not current_features:
-        # This case might occur if the single layer_item was an empty FeatureCollection or invalid
-        print(
-            f"Warning: The provided layer item is empty or not a recognizable Feature/FeatureCollection: {type(layer_item)}"
-        )
-        return []
-
-    try:
-        gdf = gpd.GeoDataFrame.from_features(current_features)
-        gdf.set_crs("EPSG:4326", inplace=True)
-
-        gdf_reprojected = gdf.to_crs(buffer_crs)
-        gdf_reprojected["geometry"] = gdf_reprojected.geometry.buffer(
-            actual_radius_meters
-        )
-        gdf_buffered_individual = gdf_reprojected.to_crs("EPSG:4326")
-
-        if gdf_buffered_individual.empty:
-            return []  # Resulting GeoDataFrame is empty
-
-        fc = json.loads(gdf_buffered_individual.to_json())
-        return [fc]  # Return a list containing the single FeatureCollection
-    except Exception as e:
-        logger.exception(f"Error in op_buffer: {e}")
-
-
-def op_centroid(layers: List[Dict[str, Any]], **kwargs) -> List[Dict[str, Any]]:
-    """
-    Compute the centroid of each feature in the first FeatureCollection and
-    return a new FeatureCollection of Point features.
-    """
-    if not layers:
-        return []
-    layer = layers[0]
-    feats = layer.get("features", [])
-    if not feats:
-        return []
-
-    try:
-        gdf = gpd.GeoDataFrame.from_features(feats)
-        gdf.set_crs("EPSG:4326", inplace=True)
-        centroids = gdf
-        centroids["geometry"] = gdf.geometry.centroid
-        fc = json.loads(centroids.to_json())
-        return [fc]
-    except Exception as e:
-        logger.exception(f"Error in op_centroid: {e}")
-        return []
-
-
-def op_simplify(
-    layers: List[Dict[str, Any]],
-    tolerance: float = 0.01,
-    preserve_topology: bool = True,
-) -> List[Dict[str, Any]]:
-    """
-    Simplify each feature in the first FeatureCollection with the given tolerance.
-    - tolerance: distance parameter for simplification (in the layer's CRS units, usually degrees).
-    - preserve_topology: whether to preserve topology during simplification.
-    """
-    feats = _flatten_features(layers)
-    if not feats:
-        return []
-    try:
-        gdf = gpd.GeoDataFrame.from_features(feats)
-        gdf["geometry"] = gdf.geometry.simplify(
-            tolerance, preserve_topology=preserve_topology
-        )
-        fc = json.loads(gdf.to_json())
-        return [fc]
-    except Exception as e:
-        logger.exception(f"Error in op_simplify: {e}")
-        return []
-
-
-def op_overlay(
-    layers: List[Dict[str, Any]], how: str = "intersection", crs: str = "EPSG:3857"
-) -> List[Dict[str, Any]]:
-    """
-    Perform a set-based overlay across N layers. Supports 'intersection', 'union',
-    'difference', 'symmetric_difference', and 'identity'. For N > 2, applies the operation iteratively:
-      result = overlay(layer1, layer2, how)
-      result = overlay(result, layer3, how)
-      crs : str, default "EPSG:3857"
-        Working Coordinate Reference System used *internally* for the overlay.
-        Provide a projected CRS (e.g. equal-area or Web Mercator) for more
-        accurate operations. **The output will always be reprojected to
-        EPSG:4326.**
-      ...
-    Returns a single FeatureCollection (in a list) of the final result.
-    """
-    if len(layers) < 2:
-        # Not enough layers to overlay; return original layers unchanged
-        return layers
-
-    # Helper: convert a layer to a GeoDataFrame in working CRS
-    def _layer_to_gdf(layer: Dict[str, Any]) -> gpd.GeoDataFrame:
-        feats = _flatten_features([layer])
-        gdf = gpd.GeoDataFrame.from_features(feats)
-        # Assume incoming CRS is WGS84 if undefined
-        if gdf.crs is None:
-            gdf.set_crs("EPSG:4326", inplace=True)
-        return gdf.to_crs(crs)
-
-    # Prepare the base layer
-    try:
-        result_gdf = _layer_to_gdf(layers[0])
-    except Exception as exc:
-        logger.exception("Error preparing base layer for overlay: %s", exc)
-        return []
-
-    # Iterate through the remaining layers
-    for layer in layers[1:]:
-        try:
-            next_gdf = _layer_to_gdf(layer)
-            result_gdf = gpd.overlay(result_gdf, next_gdf, how=how)
-
-            # Exit early if the intermediate result is empty
-            if result_gdf.empty:
-                return [{"type": "FeatureCollection", "features": []}]
-
-        except Exception as exc:
-            logger.exception("Error during overlay with layer: %s", exc)
-            return []
-
-    # Re‑project the final result back to EPSG:4326 for output
-    try:
-        if not result_gdf.empty:
-            result_gdf = result_gdf.to_crs("EPSG:4326")
-    except Exception as exc:
-        logger.exception("Error reprojecting result to EPSG:4326: %s", exc)
-        return [{"type": "FeatureCollection", "features": []}]
-
-    fc = json.loads(result_gdf.to_json())
-    return [fc]
-
-
-def op_merge(
-    layers: List[Dict[str, Any]], on: Optional[List[str]] = None, how: str = "inner"
-) -> List[Dict[str, Any]]:
-    """
-    Perform an attribute-based merge (join) between two layers.
-    - layers: expects exactly two FeatureCollections.
-    - on: list of column names to join on; if None, GeoPandas uses common columns.
-    - how: one of 'inner', 'left', 'right', 'outer'.
-    """
-    if len(layers) < 2:
-        return layers
-    try:
-        gdf1 = gpd.GeoDataFrame.from_features(layers[0].get("features", []))
-        gdf2 = gpd.GeoDataFrame.from_features(layers[1].get("features", []))
-        gdf1.set_crs("EPSG:4326", inplace=True)
-        gdf2.set_crs("EPSG:4326", inplace=True)
-
-        merged = gdf1.merge(gdf2.drop(columns="geometry"), on=on, how=how)
-        # Retain geometry from gdf1
-        merged.set_geometry(gdf1.geometry.name, inplace=True)
-        return [json.loads(merged.to_json())]
-    except Exception as e:
-        logger.exception(f"Error in op_merge: {e}")
-        return []
-
-
-def op_sjoin(
-    layers: List[Dict[str, Any]], how: str = "inner", predicate: str = "intersects"
-) -> List[Dict[str, Any]]:
-    """
-    Perform a spatial join between two layers.
-    - layers: expects exactly two FeatureCollections (left, right).
-    - how: 'left', 'right', or 'inner'.
-    - predicate: spatial predicate, e.g. 'intersects', 'contains', 'within'.
-    """
-    if len(layers) < 2:
-        return layers
-    try:
-        left_gdf = gpd.GeoDataFrame.from_features(layers[0].get("features", []))
-        right_gdf = gpd.GeoDataFrame.from_features(layers[1].get("features", []))
-        left_gdf.set_crs("EPSG:4326", inplace=True)
-        right_gdf.set_crs("EPSG:4326", inplace=True)
-
-        joined = gpd.sjoin(left_gdf, right_gdf, how=how, predicate=predicate)
-        return [json.loads(joined.to_json())]
-    except Exception as e:
-        logger.exception(f"Error in op_sjoin: {e}")
-        return []
-
-
-def op_sjoin_nearest(
-    layers: List[Dict[str, Any]],
-    how: str = "inner",
-    max_distance: Optional[float] = None,
-    distance_col: Optional[str] = None,
-) -> List[Dict[str, Any]]:
-    """
-    Perform a nearest-neighbor spatial join between two layers.
-    - layers: expects exactly two FeatureCollections (left, right).
-    - how: 'left', 'right', or 'inner'.
-    - max_distance: maximum search radius (in layer CRS units).
-    - distance_col: name of the output column to store distance.
-    """
-    if len(layers) < 2:
-        return layers
-    try:
-        left_gdf = gpd.GeoDataFrame.from_features(layers[0].get("features", []))
-        right_gdf = gpd.GeoDataFrame.from_features(layers[1].get("features", []))
-        left_gdf.set_crs("EPSG:4326", inplace=True)
-        right_gdf.set_crs("EPSG:4326", inplace=True)
-
-        joined = gpd.sjoin_nearest(
-            left_gdf,
-            right_gdf,
-            how=how,
-            max_distance=max_distance,
-            distance_col=distance_col,
-        )
-        return [json.loads(joined.to_json())]
-    except Exception as e:
-        logger.exception(f"Error in op_sjoin_nearest: {e}")
-        return []
-
 
 # Registry of available tools
 TOOL_REGISTRY = {
