+++ backend/services/tools/geoprocessing/__init__.py
@@ -1,0 +1,3 @@
+"""
+Package for geoprocessing operations and utilities.
+"""
